fix: Address PR feedback and add tests
