+++ src/services/MovieService/index.ts
@@ -1,2 +1,1 @@
 export * from './MovieService';
-export * from './types';
